@@ -6,7 +6,7 @@
     "homepage": "https://github.com/rlanvin/php-rrule",
     "license": "MIT",
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.6.0"
     },
     "suggest": {
         "ext-intl": "Intl extension is needed for humanReadable()"
@@ -16,7 +16,12 @@
             "RRule\\": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "RRule\\Tests\\": "tests/"
+        }
+    },
     "require-dev": {
-        "phpunit/phpunit": "<6"
+        "phpunit/phpunit": "^4.8|^5.5|^6.5"
     }
 }
